Use responder events for scroll and swipe.
The swipe component is a new component that takes
a side component that is displayed when
this.props.children is swiped.

I'm not happy with it, but it does the job for now

Also, I did not figure out how to properly use the
responder event, so the solution is a bit hackish.
We only use the responder chain for the move and 
end events, but use touch start since both scroll
and swipe need to respond to it. The solution
would be to have a transparent view on top of
the scroll and swipe components, but I don't feel
like doing it (nor do I have the time).

Finally, we only respond to touch events. I had a
problem when responding to all move events, and
did not take the time to check if it's still the
case.

If anyone is interested in working on making those
components better, feel free to do so, I'll be
more than happy to give you a hand.